File gfx/background.png now contains background for walls and snake body parts. It also contains the tile for the apple.
Removed aseprite files.
Added .tilemap file for the background instead, which can be edited
with tilemap-studio.

Added function to spawn a new apple after eating one. It spawns on a
random tile that is empty.

Added a call to wait for vblank before moving the head. If the snake
gets too long it took too much vblank time so it did not move the head.